requirements.txt
discord.py==2.5.2
gTTS==2.5.4
requests==2.32.4
PyNaCl==1.5.0
python-dotenv==1.0.1
